[New] Experimental support for 14-bit MIDI controllers: When recording parameter automation through MIDI mapping, combinations of CC 0...31 and CC 32...63 now act as double-precision MIDI messages (I have no devices which send such CC combinations, hence it's untested)
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5159 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Moddoc.cpp
@@ -981,7 +981,7 @@ void CModDoc::ProcessMIDI(uint32 midiData, INSTRUMENTINDEX ins, IMixPlugin *plug
 
 	PLUGINDEX mappedIndex = 0;
 	PlugParamIndex paramIndex = 0;
-	uint8 paramValue = 0;
+	uint16 paramValue = 0;
 	bool captured = m_SndFile.GetMIDIMapper().OnMIDImsg(midiData, mappedIndex, paramIndex, paramValue);
 
 	// Handle MIDI messages assigned to shortcuts
